+++ mystash-service/src/main/java/org/ranjangeorge/mystash/service/impl/ledger/FetchBalance.java
@@ -7,17 +7,19 @@ import org.ranjangeorge.mystash.service.api.support.UsecaseNames;
 import org.ranjangeorge.mystash.service.impl.Stash;
 import org.ranjangeorge.mystash.service.impl.support.db.SessionFactoryHolder;
 
+import java.math.BigDecimal;
+
 @UsecaseNames(Usecase.FETCH_BALANCE)
 public class FetchBalance {
 
-    public double fetchBalance(String stashId) {
+    public BigDecimal fetchBalance(String stashId) {
 
         Session session = SessionFactoryHolder.getCurrentSession();
         Transaction transaction = session.beginTransaction();
 
         try {
 
-            double balance = session.load(Stash.class, stashId).getBalance();
+            BigDecimal balance = session.load(Stash.class, stashId).getBalance();
 
             // Commit
             transaction.commit();
